// frontend/src/app/core/api/generated/model/patchedSmileitCatalogEntryCreateRequest.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Payload para crear sustituyente en catálogo persistente.
 */
export interface PatchedSmileitCatalogEntryCreateRequest { 
    name?: string;
    smiles?: string;
    anchor_atom_indices?: Array<number>;
    category_keys?: Array<string>;
    source_reference?: string;
    provenance_metadata?: { [key: string]: string; };
}