// frontend/src/app/core/api/generated/model/smileitAssignmentBlockInputRequest.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SmileitSubstituentReferenceInputRequest } from './smileitSubstituentReferenceInputRequest';
import { SmileitManualSubstituentInputRequest } from './smileitManualSubstituentInputRequest';


/**
 * Bloque de asignación de sustituyentes a uno o más sitios.
 */
export interface SmileitAssignmentBlockInputRequest { 
    label: string;
    site_atom_indices: Array<number>;
    category_keys?: Array<string>;
    substituent_refs?: Array<SmileitSubstituentReferenceInputRequest>;
    manual_substituents?: Array<SmileitManualSubstituentInputRequest>;
}