// frontend/src/app/core/api/generated/model/smileitPatternEntryCreateRequest.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PatternTypeEnum } from './patternTypeEnum';


/**
 * Payload para crear patrón estructural con caption obligatorio.
 */
export interface SmileitPatternEntryCreateRequest { 
    name: string;
    smarts: string;
    pattern_type: PatternTypeEnum;
    caption: string;
    source_reference?: string;
    provenance_metadata?: { [key: string]: string; };
}
export namespace SmileitPatternEntryCreateRequest {
}